Core bot is fully working. Fixes and improvements listed below.
Much better placement of the decoding operation on IRC messages.

Fixed KICK, PART, and INVITE responses and added console messages.

Fixed the "we are" message to only trigger once.

'admin_salt' is now a *bytes* object.

Changed the NickServ identification to a raw server message.

Restored any prior uses of str.format() to the % operator;
format() is slower and unneeded for the purposes I had for it.

Updated the example configuration.